% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buffer_stats}
\alias{buffer_stats}
\title{Buffer Stats S3 Class Constructor}
\usage{
buffer_stats(df)
}
\arguments{
\item{df}{data.frame of buffer stats}
}
\value{
An object of class \code{buffer_stats}
}
\description{
Buffer Stats S3 Class Constructor
}
