% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_buffer_stats}
\alias{calculate_buffer_stats}
\title{Calculate Buffer Statistics}
\usage{
calculate_buffer_stats(analysis_result)
}
\arguments{
\item{analysis_result}{Output from wind_suitability_analysis()}
}
\value{
buffer_stats S3 object with buffer contribution statistics
}
\description{
Analyze the contribution of different buffer types to total exclusions.
}
