% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_suitability_results-methods.R
\name{district_names}
\alias{district_names}
\title{Get district names from wind_suitability_results}
\usage{
district_names(x, ...)
}
\arguments{
\item{x}{A wind_suitability_results object.}

\item{...}{Ignored.}
}
\value{
Character vector of district names (unique, in order in data)
}
\description{
Get district names from wind_suitability_results
}
