% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_report.R
\name{exclusion_summary}
\alias{exclusion_summary}
\title{Exclusion summary for district_report}
\usage{
exclusion_summary(x, ...)
}
\arguments{
\item{x}{A \code{district_report} object}

\item{...}{Unused}
}
\description{
Prints exclusion and suitability statistics for the district.
}
