% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_results}
\alias{export_results}
\title{Export Analysis Results}
\usage{
export_results(analysis_result, output_dir = ".", formats = c("csv", "gpkg"))
}
\arguments{
\item{analysis_result}{Output from wind_suitability_analysis()}

\item{output_dir}{Directory to save files}

\item{formats}{Vector of formats: "csv", "gpkg", "html"}
}
\description{
Save analysis results to various formats for further use.
}
