% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.buffer_stats}
\alias{plot.buffer_stats}
\title{Plot method for buffer_stats (simple bar chart)}
\usage{
\method{plot}{buffer_stats}(x, ...)
}
\arguments{
\item{x}{buffer_stats object}

\item{...}{further arguments}
}
\description{
Plot method for buffer_stats (simple bar chart)
}
