% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_report.R
\name{plot.district_report}
\alias{plot.district_report}
\title{Plot method for district_report}
\usage{
\method{plot}{district_report}(x, ...)
}
\arguments{
\item{x}{A \code{district_report} object}

\item{...}{Unused}
}
\description{
Displays a bar plot of suitable and excluded area percentages for the district.
}
