% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_suitability_results-methods.R
\name{plot.wind_suitability_results}
\alias{plot.wind_suitability_results}
\title{Plot method for wind_suitability_results}
\usage{
\method{plot}{wind_suitability_results}(x, type = "bar", ...)
}
\arguments{
\item{x}{A wind_suitability_results object.}

\item{type}{Plot type: "bar" or "map".}

\item{...}{Additional arguments (currently unused).}
}
\description{
Plot method for wind_suitability_results
}
