% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.buffer_stats}
\alias{print.buffer_stats}
\title{Print method for buffer_stats}
\usage{
\method{print}{buffer_stats}(x, ...)
}
\arguments{
\item{x}{buffer_stats object}

\item{...}{further arguments}
}
\description{
Print method for buffer_stats
}
