% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_report.R
\name{print.district_report}
\alias{print.district_report}
\title{Print method for district_report}
\usage{
\method{print}{district_report}(x, ...)
}
\arguments{
\item{x}{A \code{district_report} object}

\item{...}{Unused}
}
\description{
Print method for district_report
}
