% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.buffer_stats}
\alias{summary.buffer_stats}
\title{Summary method for buffer_stats}
\usage{
\method{summary}{buffer_stats}(object, ...)
}
\arguments{
\item{object}{buffer_stats object}

\item{...}{further arguments}
}
\description{
Summary method for buffer_stats
}
