% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_suitability_results-methods.R
\name{summary.wind_suitability_results}
\alias{summary.wind_suitability_results}
\title{Summary method for wind_suitability_results}
\usage{
\method{summary}{wind_suitability_results}(object, ...)
}
\arguments{
\item{object}{A wind_suitability_results object.}

\item{...}{Ignored.}
}
\description{
Summary method for wind_suitability_results
}
