% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district_report.R
\name{to_geojson.district_report}
\alias{to_geojson.district_report}
\title{Export district geometry to GeoJSON}
\usage{
\method{to_geojson}{district_report}(x, file = paste0(x$name, "_district.geojson"), ...)
}
\arguments{
\item{x}{A \code{district_report} object}

\item{file}{Filename for the GeoJSON output}

\item{...}{Unused}
}
\description{
Writes the district's geometry to a GeoJSON file.
}
