% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_suitability_results-methods.R
\name{top_districts}
\alias{top_districts}
\title{Get top districts by suitability}
\usage{
top_districts(x, n = 5, ...)
}
\arguments{
\item{x}{A wind_suitability_results object.}

\item{n}{Number of top districts to return (default: 5)}

\item{...}{Ignored.}
}
\value{
Data frame of the top N districts by suitability.
}
\description{
Get top districts by suitability
}
