% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R, R/wind_suitability_results-methods.R
\name{print.wind_suitability_results}
\alias{print.wind_suitability_results}
\title{Print method for wind_suitability_results}
\usage{
\method{print}{wind_suitability_results}(x, ...)

\method{print}{wind_suitability_results}(x, ...)
}
\arguments{
\item{x}{A wind_suitability_results object.}

\item{...}{Ignored.}
}
\description{
Print method for wind_suitability_results

Print method for wind_suitability_results
}
